requirements.txt
git+https://github.com/AI4Finance-Foundation/FinRL.git
stable-baselines3==2.2.1
gymnasium==0.29.1
pandas==1.5.3
numpy==1.24.3
matplotlib==3.7.1
alpaca-py>=0.37,<0.38
python-dotenv==1.0.0
scikit-learn==1.2.2
torch==2.0.1
pandas_market_calendars
